requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
deepgram-sdk==3.1.4
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
verboselogs==1.7
websockets==12.0
yarl==1.9.4
